モーダルの中身、レイアウトを修正
