+++ work/resources/ts/containers/organisms/DragBoardList.tsx
@@ -3,6 +3,7 @@ import { DragDropContext, Droppable } from 'react-beautiful-dnd';
 import { DragBoardItem } from '../../components/molecules/DragBoardItem';
 import { AddIcon } from '../../components/atoms/AddIcon';
 import { ModalWindow } from '../../components/molecules/ModalWindow';
+import { ModalCard } from '../../components/molecules/ModalCard';
 import { StylesContext } from '../../contexts/childContexts/StylesContext';
 import { BoardItemContext } from '../../contexts/childContexts/BoardItemContext';
 import { ShowCardsContext } from '../../contexts/childContexts/ShowCardsContext';
@@ -101,48 +102,53 @@ const regularExpressions = /^.+/;
         />
       </div>
       <ModalWindow
-        errorTitle={regularExpressions.test(modalValueState.title) ? false : true}
-        helperTextTitle={
-          regularExpressions.test(modalValueState.title) ? undefined : 'タイトルを入力してください'
-        }
-        errorContent={regularExpressions.test(modalValueState.content) ? false : true}
-        helperTextContent={
-          regularExpressions.test(modalValueState.content) ? undefined : '内容を入力してください'
-        }
         modalOpen={modalOpenState}
         modalOnClose={modalClose}
-        // 押したボタンの番号によって、表示内容を変える
-        defaultValueTitle={modalValueState.title}
-        defaultValueContent={modalValueState.content}
-        titleOnChange={(e) => {
-          setmodalValueState({ ...modalValueState, title: e.target.value })
-        }}
-        contentOnChange={(e) => {
-          setmodalValueState({ ...modalValueState, content: e.target.value })
-        }}
-        postOnClick={() => {
-          /** 今のBoardItemの配列を受け取り、更新部分だけ新しい値に入れ替える
-            * updateなら既存のindexに格納
-            * addならnewBoardItemState.items.lengthで最後の位置に格納
-            * indexNumberに格納位置のindexを入れる
-          */
-          const newBoardItemState = { ...cardsState };
-          const searchIndex = newBoardItemState.items.findIndex(({id}: any) => id == modalValueState.id)
-          const indexNumber = searchIndex > -1 ? searchIndex : newBoardItemState.items.length;
-          newBoardItemState.items[indexNumber] = modalValueState;
-          setCardsState(
-            newBoardItemState,
-          );
-          setCardsState(
-            { ...cardsState, numberMade: cardsState.numberMade + 1 }
-          )
-          modalClose();
-        }}
-        disabled={
-          // クソコードだから整理したい
-          modalValueState.title === null || modalValueState.content === null ? (true) : (
-          regularExpressions.test(modalValueState.title) && regularExpressions.test(modalValueState.content) ? (false) : (true))}
-      />
+      >
+        <ModalCard
+          errorTitle={regularExpressions.test(modalValueState.title) ? false : true}
+          helperTextTitle={
+            regularExpressions.test(modalValueState.title) ? undefined : 'タイトルを入力してください'
+          }
+          errorContent={regularExpressions.test(modalValueState.content) ? false : true}
+          helperTextContent={
+            regularExpressions.test(modalValueState.content) ? undefined : '内容を入力してください'
+          }
+          modalOpen={modalOpenState}
+          modalOnClose={modalClose}
+          // 押したボタンの番号によって、表示内容を変える
+          defaultValueTitle={modalValueState.title}
+          defaultValueContent={modalValueState.content}
+          titleOnChange={(e) => {
+            setmodalValueState({ ...modalValueState, title: e.target.value })
+          }}
+          contentOnChange={(e) => {
+            setmodalValueState({ ...modalValueState, content: e.target.value })
+          }}
+          postOnClick={() => {
+            /** 今のBoardItemの配列を受け取り、更新部分だけ新しい値に入れ替える
+              * updateなら既存のindexに格納
+              * addならnewBoardItemState.items.lengthで最後の位置に格納
+              * indexNumberに格納位置のindexを入れる
+            */
+            const newBoardItemState = { ...cardsState };
+            const searchIndex = newBoardItemState.items.findIndex(({id}: any) => id == modalValueState.id)
+            const indexNumber = searchIndex > -1 ? searchIndex : newBoardItemState.items.length;
+            newBoardItemState.items[indexNumber] = modalValueState;
+            setCardsState(
+              newBoardItemState,
+            );
+            setCardsState(
+              { ...cardsState, numberMade: cardsState.numberMade + 1 }
+            )
+            modalClose();
+          }}
+          disabled={
+            // クソコードだから整理したい
+            modalValueState.title === null || modalValueState.content === null ? (true) : (
+            regularExpressions.test(modalValueState.title) && regularExpressions.test(modalValueState.content) ? (false) : (true))}
+        />
+      </ModalWindow>
     </>
   )
 });
