+++ work/resources/ts/contexts/childContexts/StylesContext.tsx
@@ -23,7 +23,6 @@ export const StylesContextProvider: React.FC = props => {
     modal: {
       position: 'absolute',
       width: `60%`,
-      height: `60%`,
       backgroundColor: theme.palette.background.paper,
       border: '2px solid #000',
       boxShadow: theme.shadows[5],
